Move shared smartcard related data to own file.
