+++ src/__tests__/helpers/smartcards.tsx
@@ -1,0 +1,60 @@
+import GLOBALS from '../../config/globals'
+import { CardAPI, Election } from '../../config/types'
+import electionSample from '../../data/electionSample.json'
+
+const election = electionSample as Election
+
+export const noCard: CardAPI = {
+  present: false,
+}
+
+export const adminCard: CardAPI = {
+  present: true,
+  longValueExists: true,
+  shortValue: JSON.stringify({
+    t: 'clerk',
+    h: 'abcd',
+  }),
+}
+
+export const pollWorkerCard: CardAPI = {
+  present: true,
+  shortValue: JSON.stringify({
+    t: 'pollworker',
+    h: 'abcd',
+  }),
+}
+
+const voterCardShortValue = {
+  t: 'voter',
+  pr: election.precincts[0].id,
+  bs: election.ballotStyles[0].id,
+}
+
+export const voterCard: CardAPI = {
+  present: true,
+  shortValue: JSON.stringify(voterCardShortValue),
+}
+
+export const invalidatedVoterCard: CardAPI = {
+  present: true,
+  shortValue: JSON.stringify({
+    ...voterCardShortValue,
+    uz: new Date().getTime(),
+  }),
+}
+
+export const getPrintedVoterCard = (): CardAPI => ({
+  present: true,
+  shortValue: JSON.stringify({
+    ...voterCardShortValue,
+    bp: 1,
+    uz: new Date().getTime(),
+  }),
+})
+
+export const advanceTimers = (ms: number = 0) => {
+  jest.advanceTimersByTime(ms + GLOBALS.CARD_POLLING_INTERVAL)
+}
+
+export default {}
